requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pymongo==3.12.0
python-dotenv=0.19.0
mangum==0.12.0
pydantic
pydantic-settings
dotenv
farm-haystack
awscli
awsebcli
